pyproject.toml
[tool.poetry]
name = "myproject"
version = "0.0.1"
description = ""
authors = ["myproject <myproject@example.com>"]
readme = "README.md"
packages = [{include = "myproject"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
pytest-cov = "^4.1.0"
commitizen = "^3.6.0"
poetry2conda = "^0.3.0"
pylint = "^2.17.5"
pdoc3 = "^0.10.0"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry2conda]
name = "myproject-env"

[tool.poetry2conda.dependencies]
foo = { channel = "conda-forge" }
